// prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  PIC: 'PIC'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const BookingStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const BookingSession = {
  SESSION_1: 'SESSION_1',
  SESSION_2: 'SESSION_2',
  FULLDAY: 'FULLDAY'
} as const

export type BookingSession = (typeof BookingSession)[keyof typeof BookingSession]


export const MeetingType = {
  INTERNAL: 'INTERNAL',
  INTERNAL_LINTAS_BIDANG: 'INTERNAL_LINTAS_BIDANG',
  EKSTERNAL: 'EKSTERNAL'
} as const

export type MeetingType = (typeof MeetingType)[keyof typeof MeetingType]